Renaming and refactoring analyser classes
