+++ src/DataPrivacyAuditTool.Tests.Unit/Analysers/BaseAnalyserTests.cs
@@ -1,6 +1,6 @@
 using DataPrivacyAuditTool.Core.Interfaces;
 using DataPrivacyAuditTool.Core.Models;
-using DataPrivacyAuditTool.Infrastructure.Services.Analyzers;
+using DataPrivacyAuditTool.Infrastructure.Services.Analysers;
 using System;
 using System.Threading.Tasks;
 using Xunit;
@@ -13,12 +13,12 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
     public class BaseAnalyzersTests
     {
         // Test implementations of abstract classes
-        private class TestSettingsAnalyzer:SettingsAnalyzer
+        private class TestSettingsAnalyser:SettingsAnalyser
         {
             public override string CategoryName => "Test Settings Category";
             public override string Description => "Test Settings Description";
 
-            protected override Task<PrivacyMetricCategory> AnalyzeSettingsAsync(SettingsData settingsData)
+            protected override Task<PrivacyMetricCategory> AnalyseSettingsAsync(SettingsData settingsData)
             {
                 return Task.FromResult(new PrivacyMetricCategory
                 {
@@ -29,12 +29,12 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
             }
         }
 
-        private class TestAddressesAnalyzer:AddressesAnalyzer
+        private class TestAddressesAnalyser:AddressesAnalyser
         {
             public override string CategoryName => "Test Addresses Category";
             public override string Description => "Test Addresses Description";
 
-            protected override Task<PrivacyMetricCategory> AnalyzeAddressesAsync(AddressData addressesData)
+            protected override Task<PrivacyMetricCategory> AnalyseAddressesAsync(AddressData addressesData)
             {
                 return Task.FromResult(new PrivacyMetricCategory
                 {
@@ -49,14 +49,14 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
         public async Task SettingsAnalyzer_CallsAnalyzeSettingsAsync_WhenDataIsProvided()
         {
             // Arrange
-            var analyzer = new TestSettingsAnalyzer();
+            var analyzer = new TestSettingsAnalyser();
             var parsedData = new ParsedGoogleData
             {
                 SettingsData = new SettingsData()
             };
 
             // Act
-            var result = await analyzer.AnalyzeAsync(parsedData);
+            var result = await analyzer.AnalyseAsync(parsedData);
 
             // Assert
             Assert.NotNull(result);
@@ -69,7 +69,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
         public async Task SettingsAnalyzer_ThrowsArgumentException_WhenSettingsDataIsNull()
         {
             // Arrange
-            var analyzer = new TestSettingsAnalyzer();
+            var analyzer = new TestSettingsAnalyser();
             var parsedData = new ParsedGoogleData
             {
                 SettingsData = null,
@@ -77,21 +77,21 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
             };
 
             // Act & Assert
-            await Assert.ThrowsAsync<ArgumentException>(() => analyzer.AnalyzeAsync(parsedData));
+            await Assert.ThrowsAsync<ArgumentException>(() => analyzer.AnalyseAsync(parsedData));
         }
 
         [Fact]
         public async Task AddressesAnalyzer_CallsAnalyzeAddressesAsync_WhenDataIsProvided()
         {
             // Arrange
-            var analyzer = new TestAddressesAnalyzer();
+            var analyzer = new TestAddressesAnalyser();
             var parsedData = new ParsedGoogleData
             {
                 AddressesData = new AddressData()
             };
 
             // Act
-            var result = await analyzer.AnalyzeAsync(parsedData);
+            var result = await analyzer.AnalyseAsync(parsedData);
 
             // Assert
             Assert.NotNull(result);
@@ -104,7 +104,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
         public async Task AddressesAnalyzer_ThrowsArgumentException_WhenAddressesDataIsNull()
         {
             // Arrange
-            var analyzer = new TestAddressesAnalyzer();
+            var analyzer = new TestAddressesAnalyser();
             var parsedData = new ParsedGoogleData
             {
                 SettingsData = new SettingsData(),
@@ -112,7 +112,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
             };
 
             // Act & Assert
-            await Assert.ThrowsAsync<ArgumentException>(() => analyzer.AnalyzeAsync(parsedData));
+            await Assert.ThrowsAsync<ArgumentException>(() => analyzer.AnalyseAsync(parsedData));
         }
     }
 }
