+++ src/DataPrivacyAuditTool.Tests.Unit/Analysers/PersonalDataExposureAnalyserTests.cs
@@ -1,5 +1,5 @@
 using DataPrivacyAuditTool.Core.Models;
-using DataPrivacyAuditTool.Infrastructure.Services.Analyzers;
+using DataPrivacyAuditTool.Infrastructure.Services.Analysers;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Xunit;
@@ -12,15 +12,15 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
     /// </summary>
     public class PersonalDataExposureAnalyserTests
     {
-        private readonly PersonalDataExposureAnalyser _analyzer;
+        private readonly PersonalDataExposureAnalyser _analyser;
 
         public PersonalDataExposureAnalyserTests()
         {
-            _analyzer = new PersonalDataExposureAnalyser();
+            _analyser = new PersonalDataExposureAnalyser();
         }
 
         [Fact]
-        public async Task AnalyzeAsync_EmailExposure_DetectsMultipleEmails()
+        public async Task AnalyseAsync_EmailExposure_DetectsMultipleEmails()
         {
             // Arrange
             var parsedData = new ParsedGoogleData
@@ -51,7 +51,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
             };
 
             // Act
-            var result = await _analyzer.AnalyzeAsync(parsedData);
+            var result = await _analyser.AnalyseAsync(parsedData);
 
             // Assert
             Assert.NotNull(result);
@@ -64,7 +64,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
         }
 
         [Fact]
-        public async Task AnalyzeAsync_PhoneNumberExposure_DetectsStoredPhoneNumbers()
+        public async Task AnalyseAsync_PhoneNumberExposure_DetectsStoredPhoneNumbers()
         {
             // Arrange
             var parsedData = new ParsedGoogleData
@@ -90,7 +90,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
             };
 
             // Act
-            var result = await _analyzer.AnalyzeAsync(parsedData);
+            var result = await _analyser.AnalyseAsync(parsedData);
 
             // Assert
             Assert.NotNull(result);
@@ -102,7 +102,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
         }
 
         [Fact]
-        public async Task AnalyzeAsync_AddressExposure_DetectsStoredAddresses()
+        public async Task AnalyseAsync_AddressExposure_DetectsStoredAddresses()
         {
             // Arrange
             var parsedData = new ParsedGoogleData
@@ -120,7 +120,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
             };
 
             // Act
-            var result = await _analyzer.AnalyzeAsync(parsedData);
+            var result = await _analyser.AnalyseAsync(parsedData);
 
             // Assert
             Assert.NotNull(result);
@@ -133,7 +133,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
         }
 
         [Fact]
-        public async Task AnalyzeAsync_OverallExposure_CalculatesCorrectExposureLevel()
+        public async Task AnalyseAsync_OverallExposure_CalculatesCorrectExposureLevel()
         {
             // Arrange
             var parsedData = new ParsedGoogleData
@@ -167,7 +167,7 @@ namespace DataPrivacyAuditTool.Tests.Unit.Analyzers
             };
 
             // Act
-            var result = await _analyzer.AnalyzeAsync(parsedData);
+            var result = await _analyser.AnalyseAsync(parsedData);
 
             // Assert
             Assert.NotNull(result);
